[ refactor ] 불필요 interface 제거
