+++ module-service/src/main/java/module/service/genre/GenreService.java
@@ -2,8 +2,23 @@ package module.service.genre;
 
 import java.util.List;
 
-import dto.genre.GenreDTO;
+import org.modelmapper.ModelMapper;
+import org.springframework.stereotype.Service;
 
-public interface GenreService {
-	List<GenreDTO>  getAllGenre();
+import dto.genre.GenreDTO;
+import lombok.RequiredArgsConstructor;
+import module.repository.genre.GenreRepository;
+
+@Service
+@RequiredArgsConstructor
+public class GenreService {
+
+	private final GenreRepository genreRepository;
+	private final ModelMapper modelMapper;
+
+	public List<GenreDTO> getAllGenre() {
+		return genreRepository.findAll().stream()
+			.map(entity-> modelMapper.map(entity,GenreDTO.class))
+			.toList();
+	}
 }
